packages/pdf-extract/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_extract"]

[project]
name = "pdf-extract"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "anthropic",
    "openai",
    "PyMuPDF>=1.24",
    "opencv-python>=4.9",
    "Pillow>=10",
    "structlog>=24",
    "tenacity>=9",
]

[project.scripts]
pdf-extract = "pdf_extract.cli:main"

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "syrupy",
    "hypothesis",
    "rapidfuzz",
    "rouge-score",
    "fpdf2",
    "respx",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH", "RUF"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: needs real API keys or extractors installed",
    "slow: takes >10s",
]

[tool.mypy]
python_version = "3.12"
strict = true
